+++ app/list_pricing_by_sku.go
@@ -1,0 +1,30 @@
+package app
+
+import (
+	"database/sql"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"os"
+
+	"github.com/go-chi/chi"
+)
+
+// ListPricingBySKUHandler creates a handler function that returns a
+// map of tier refs to pricing entries.
+func (a *App) ListPricingBySKUHandler() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		sku := chi.URLParam(r, "sku")
+		pmap, err := a.Service.ListPricingBySKU(r.Context(), sku)
+		if err != nil {
+			if err == sql.ErrNoRows {
+				w.WriteHeader(http.StatusNotFound)
+				return
+			}
+			fmt.Fprintf(os.Stderr, "service ListPricingBySKU(ctx, %s) error: %+v", sku, err)
+			return
+		}
+		w.WriteHeader(http.StatusOK) // 200 OK
+		json.NewEncoder(w).Encode(pmap)
+	}
+}